Closed Issue #24.
Determined that binding wx.EVT_LEFT_DOWN prevents the parent panel from setting focus to the FloatCanvas, which in turn prevents any events bound to the canvas from firing. At least I think that's the case.

In the case of changing the discrete data plot color, what appears to have been happening is that the focus would move from the FloatCanvas to the Legend, and then because wx.EVT_LEFT_DOWN was bound, focus could not return to the FloatCanvas.